# docs/swagger.yaml
openapi: "3.0.3"
info:
  title: "GoGoStudy API"
  version: "1.0.0"
  description: "Auth, user, study, and attendance APIs for GoGoStudy."
servers:
  -
    url: "http://localhost:8080"
    description: "Local"
...
components:
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"
  schemas:
    ErrorResponse:
      type: "object"
      properties:
        timestamp:
          type: "string"
          format: "date-time"
        path:
          type: "string"
        status:
          type: "integer"
        code:
          type: "string"
        message:
          type: "string"
        details:
          type: "object"
          nullable: true
    User:
      type: "object"
      properties:
        id:
          type: "integer"
        email:
          type: "string"
        name:
          type: "string"
        role:
          type: "string"
        status:
          type: "string"
        provider:
          type: "string"
        providerId:
          type: "string"
          nullable: true
        createdAt:
          type: "string"
          format: "date-time"
        updatedAt:
          type: "string"
          format: "date-time"
...
    Study:
      type: "object"
      properties:
        id:
          type: "integer"
        title:
          type: "string"
        description:
          type: "string"
        category:
          type: "string"
          nullable: true
        maxMembers:
          type: "integer"
          nullable: true
        status:
          type: "string"
        leaderId:
          type: "integer"
        createdAt:
          type: "string"
          format: "date-time"
...
    AttendanceRecord:
      type: "object"
      properties:
        id:
          type: "integer"
        sessionId:
          type: "integer"
        userId:
          type: "integer"
        status:
          type: "string"
        recordedAt:
          type: "string"
          format: "date-time"
  responses:
    BadRequest:
      description: "Bad Request"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: "Unauthorized"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: "Forbidden"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: "Not Found"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalError:
      description: "Internal Server Error"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
paths:
...
  /studies:
    get:
      tags:
        - "Studies"
      summary: "List studies"
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: "q"
          in: "query"
          schema:
            type: "string"
          description: "Search by title/description"
        -
          name: "category"
          in: "query"
          schema:
            type: "string"
        -
          name: "status"
          in: "query"
          schema:
            type: "string"
        -
          name: "page"
          in: "query"
          schema:
            type: "integer"
            minimum: 1
          description: "Page number"
        -
          name: "pageSize"
          in: "query"
          schema:
            type: "integer"
            minimum: 1
            maximum: 50
          description: "Items per page"
      responses:
        200:
          description: "Studies listed"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  data:
                    type: "array"
                    items:
                      $ref: "#/components/schemas/Study"
                  page:
                    type: "integer"
                  pageSize:
                    type: "integer"
                  total:
                    type: "integer"
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        500:
          $ref: "#/components/responses/InternalError"
    post:
      tags:
        - "Studies"
      summary: "Create study"
      security:
        -
          bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: "object"
              required:
                - "title"
                - "description"
              properties:
                title:
                  type: "string"
                description:
                  type: "string"
                category:
                  type: "string"
                maxMembers:
                  type: "integer"
      responses:
        201:
          description: "Study created"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  study:
                    $ref: "#/components/schemas/Study"
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        500:
          $ref: "#/components/responses/InternalError"
...
  /admin/users/{id}/attendance:
    get:
      tags:
        - "Admin"
      summary: "Get attendance records for user"
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: "id"
          in: "path"
          required: true
          schema:
            type: "integer"
      responses:
        200:
          description: "Attendance history"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                  records:
                    type: "array"
                    items:
                      $ref: "#/components/schemas/AttendanceRecord"
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/InternalError"
  /admin/studies:
    get:
      tags:
        - "Admin"
      summary: "List studies (admin)"
      security:
        -
          bearerAuth: []
      parameters:
        -
          name: "status"
          in: "query"
          schema:
            type: "string"
        -
          name: "page"
          in: "query"
          schema:
            type: "integer"
            minimum: 1
        -
          name: "size"
          in: "query"
          schema:
            type: "integer"
            minimum: 1
            maximum: 50
      responses:
        200:
          description: "Studies listed"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  items:
                    type: "array"
                    items:
                      $ref: "#/components/schemas/Study"
                  page:
                    type: "integer"
                  size:
                    type: "integer"
                  total:
                    type: "integer"
                  totalPages:
                    type: "integer"
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        500:
          $ref: "#/components/responses/InternalError"
  /admin/stats/overview:
    get:
      tags:
        - "Admin"
      summary: "Platform overview stats"
      security:
        -
          bearerAuth: []
      responses:
        200:
          description: "Overview stats"
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  users:
                    type: "object"
                    properties:
                      total:
                        type: "integer"
                      active:
                        type: "integer"
                      admins:
                        type: "integer"
                  studies:
                    type: "object"
                    properties:
                      total:
                        type: "integer"
                      recruiting:
                        type: "integer"
                  attendance:
                    type: "object"
                    properties:
                      sessionsTotal:
                        type: "integer"
                      sessionsToday:
                        type: "integer"
                      records:
                        type: "integer"
                      pendingMembers:
                        type: "integer"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        500:
          $ref: "#/components/responses/InternalError"